Requests files on websocket open
Instead of only requesting the available files when Klippy is ready
request them when websocket opens. Moonraker handles the files, so there
is no need to wait until klippy is connected. This also ensures that
when the screen connects with moonraker and klipper the information
about all available print files are ready to be handled by the GUI.
This stems from a problem loading the thumbnails. There were many
instances of initializing the application mid print, when this
happened a request for the available thumbnails is made, because the
files are loaded after receiving the klippy ready event, the moment
when asking about the current job thumbnail there are not files nor
metadata loaded onto the GUI, so no thumbnail is fetched. Now with this
change the same situation doesn't occur.