Add inner progress bar icon
The progress bar fills a circumference until the print is complete,
the inner part of the widget is blank and is suposed to have a thumbnail
of the current print, before a QGraphicsView overlaped with the
progress bar widget. Now we add the functionality to add a pixmap on the
progress bar itself. It calculates the scaling for the pixmap to fill
the inner part of the progress bar widget without overlaping anything.
There are some bugs still, such as what happens when resizeEvent is
emitted, the automatic scaling is not working properly yet.