Refactor Thumnail painting and widget building
The solution provided by @Robert0Mart showed us that the thumbnail
building needed some work, this commit intends to extend that effort to
rewrite this feature. We will simplify thumbnail building, less
widgets for this feature to work. We will also let the progress bar
widget handle the small widget painting, and propagate the click to
expand to a bigger thumbnail painting that fills the entire screen.