Final progress bar refactor, send signal on thumbnail click
Refactored class method positions to add some structure to the class.
Privatised some class variables, should add @property decorator with
getter and setter for these variables since they are actually properties.
Captured MousePressEvent of the widget and filtered the events only to
those inside the inner rect of the widget, when triggered, send an event
`thumbnail_clicked` to signal that the thumbnail was actually clicked.