Added thumbnail expansion
Added initial thumbnail expansion, it works, but it needs to be worked
on. We need to hide all widgets so that the thumbnail doesn't have anything
on the background while its expanded. this requires us to hide all widgets
used one by one. it's not incorrect per-se but it's not the best. The thumbnail
is loaded once when the thumbnail is received, so it just paints once and stays
in memory until `.show()` and `.hide()` methods are called. Once the print stops
for whatever reason. Since the widget is built when the thumbnail comes, the
QGraphicsView and its childs are deleted, freeing memory. It'll be built again
when a new print job is selected.