Refactor filter null pixmaps during thumbnail loading
Simplified thumbnail filtering. Now i can have any number of Null
pixmaps, if no pixmap is usable, cancels thumbnail loading,
if there is at least one usable pixmap. It'll load the thumbnails
on the progress widget and on the QGraphicsView.
This means that if only the smallest resolution is available (48x48)
it'll paint with low resolution