Del: reference to uninitialized QGraphicsView variable
When there is no thumbnail for the current print, we shouldn't referece
`self.thumbnail_view` since it hasn't been initialized. When the print
stops for whatever reason we want to delete the object, but if it hasn't
been initilialized there is nothing to delete. So we must only delete
when the print stops if the class has been attributed.