Calculate and scale thumbnail pixmap on set
When setting the pixmap on the progress bar, the image was not scaled and
the inner rect was not calculated. This resulted in the pixmap not
showing. Now when setting the progressbar thumbnail this is calculated
so that the pixmap can be shown in the middle of the progress bar
circumference.