Refactor and handle show event
Refactored some methods, including accessing values in dicts,
by using .get().
Now the slot `on_fileinfo` only runs when the `jobStatusPage` is visible
this is because the request for file information is done on `file.py` and
the confirmation page. This whould result in the slot triggering multiple
times before it was actually necessary and on asking for imformation for
all files, while we only want information on one file. Now the request
for file information is done when the `jobStatusPage` is actually
visible. The `showEvent` method requests the file information when that
event is triggered on the class.